Refactors the app structure
There is two new folders :
- `data` for data querying and processing logic;
- `layout` for HTML layout;

The modules inside these folders are named public or internal depending on
the data they process or serve.
