+++ app/data/public.py
@@ -10,12 +10,9 @@ from zoneinfo import ZoneInfo
 import pandas as pd
 import sqlalchemy
 
-from app.cache_config import appcache
-from app.time_config import *
 
-# postgresql://admin:admin@localhost:5432/ibnse
 DB_ENGINE = sqlalchemy.create_engine(getenv("DATABASE_URL"))
-SQL_PATH = Path("app/sql")
+SQL_PATH = Path(__file__).parent.absolute() / "sql"
 
 
 def get_bsdd_data() -> pd.DataFrame:
